+++ services/telemetry/telemetry-linux/test/msg.c
@@ -52,7 +52,7 @@ static void test_subscribe_msg(void ** arg)
 {
     uint8_t buffer[100];
     telemetry_message_type msg_type;
-    int topic_in = 10, topic_out;
+    uint16_t topic_in = 10, topic_out;
 
     int msg_size = telemetry_encode_subscribe_msg(buffer, &topic_in);
     bool parsed_type = telemetry_parse_msg_type(buffer, msg_size, &msg_type);
@@ -70,7 +70,7 @@ static void test_unsubscribe_msg(void ** arg)
 {
     uint8_t buffer[100];
     telemetry_message_type msg_type;
-    int topic_in = 11, topic_out;
+    uint16_t topic_in = 11, topic_out;
 
     int msg_size = telemetry_encode_unsubscribe_msg(buffer, &topic_in);
     bool parsed_type = telemetry_parse_msg_type(buffer, msg_size, &msg_type);
